# Improve performance of FastOR