@lemire 2018-08-20
Using go 1.10 on a skylake processor. 

Numbers before pull this PR: (two runs)

```
$ BENCH_REAL_DATA=1 go test -bench BenchmarkRealDataFastOr -run -
goos: linux
goarch: amd64
pkg: github.com/RoaringBitmap/roaring
BenchmarkRealDataFastOr/census-income_srt         	    3000	    416473 ns/op
BenchmarkRealDataFastOr/census-income             	    2000	    865373 ns/op
BenchmarkRealDataFastOr/census1881_srt            	    2000	   1106882 ns/op
BenchmarkRealDataFastOr/census1881                	    1000	   1727290 ns/op
BenchmarkRealDataFastOr/dimension_003             	    1000	   1812480 ns/op
BenchmarkRealDataFastOr/dimension_008             	    2000	    820555 ns/op
BenchmarkRealDataFastOr/dimension_033             	    2000	    648946 ns/op
BenchmarkRealDataFastOr/uscensus2000              	     500	   3088391 ns/op
BenchmarkRealDataFastOr/weather_sept_85_srt       	    5000	    358800 ns/op
BenchmarkRealDataFastOr/weather_sept_85           	     500	   3873030 ns/op
BenchmarkRealDataFastOr/wikileaks-noquotes_srt    	   10000	    226216 ns/op
BenchmarkRealDataFastOr/wikileaks-noquotes        	    5000	    372601 ns/op
PASS
ok  	github.com/RoaringBitmap/roaring	49.104s

$ BENCH_REAL_DATA=1 go test -bench BenchmarkRealDataFastOr -run -
goos: linux
goarch: amd64
pkg: github.com/RoaringBitmap/roaring
BenchmarkRealDataFastOr/census-income_srt         	    3000	    416055 ns/op
BenchmarkRealDataFastOr/census-income             	    2000	    864309 ns/op
BenchmarkRealDataFastOr/census1881_srt            	    2000	   1101729 ns/op
BenchmarkRealDataFastOr/census1881                	    1000	   1725244 ns/op
BenchmarkRealDataFastOr/dimension_003             	    1000	   1680860 ns/op
BenchmarkRealDataFastOr/dimension_008             	    2000	    820634 ns/op
BenchmarkRealDataFastOr/dimension_033             	    2000	    659835 ns/op
BenchmarkRealDataFastOr/uscensus2000              	     500	   3114915 ns/op
BenchmarkRealDataFastOr/weather_sept_85_srt       	    5000	    353486 ns/op
BenchmarkRealDataFastOr/weather_sept_85           	     500	   3872940 ns/op
BenchmarkRealDataFastOr/wikileaks-noquotes_srt    	   10000	    226217 ns/op
BenchmarkRealDataFastOr/wikileaks-noquotes        	    5000	    372520 ns/op
PASS
ok  	github.com/RoaringBitmap/roaring	48.979s
```

Numbers after this PR, two runs:

```
$ BENCH_REAL_DATA=1 go test -bench BenchmarkRealDataFastOr -run -
goos: linux
goarch: amd64
BenchmarkRealDataFastOr/census-income_srt         	    3000	    404340 ns/op
BenchmarkRealDataFastOr/census-income             	    2000	    773451 ns/op
BenchmarkRealDataFastOr/census1881_srt            	    2000	   1098004 ns/op
BenchmarkRealDataFastOr/census1881                	    1000	   1446121 ns/op
BenchmarkRealDataFastOr/dimension_003             	    1000	   1700979 ns/op
BenchmarkRealDataFastOr/dimension_008             	    2000	    836829 ns/op
BenchmarkRealDataFastOr/dimension_033             	    2000	    646906 ns/op
BenchmarkRealDataFastOr/uscensus2000              	     500	   3091823 ns/op
BenchmarkRealDataFastOr/weather_sept_85_srt       	    5000	    354086 ns/op
BenchmarkRealDataFastOr/weather_sept_85           	     500	   3362478 ns/op
BenchmarkRealDataFastOr/wikileaks-noquotes_srt    	   10000	    229325 ns/op
BenchmarkRealDataFastOr/wikileaks-noquotes        	    5000	    372595 ns/op
PASS
ok  	_/home/dlemire/tmp/roaring	48.194s
dlemire@skylake:~/tmp/roaring$ BENCH_REAL_DATA=1 go test -bench BenchmarkRealDataFastOr -run -
goos: linux
goarch: amd64
BenchmarkRealDataFastOr/census-income_srt         	    5000	    404096 ns/op
BenchmarkRealDataFastOr/census-income             	    2000	    776094 ns/op
BenchmarkRealDataFastOr/census1881_srt            	    2000	   1097688 ns/op
BenchmarkRealDataFastOr/census1881                	    1000	   1448524 ns/op
BenchmarkRealDataFastOr/dimension_003             	    1000	   1689910 ns/op
BenchmarkRealDataFastOr/dimension_008             	    2000	    774278 ns/op
BenchmarkRealDataFastOr/dimension_033             	    2000	    659762 ns/op
BenchmarkRealDataFastOr/uscensus2000              	     500	   3063117 ns/op
BenchmarkRealDataFastOr/weather_sept_85_srt       	    5000	    353139 ns/op
BenchmarkRealDataFastOr/weather_sept_85           	     500	   3370650 ns/op
BenchmarkRealDataFastOr/wikileaks-noquotes_srt    	   10000	    227178 ns/op
BenchmarkRealDataFastOr/wikileaks-noquotes        	    5000	    372051 ns/op
PASS
ok  	_/home/dlemire/tmp/roaring	48.884s```


This PR is clearly beneficial some of the time.