@lemire 2018-08-20
You have been added as a contributor:

https://github.com/RoaringBitmap/roaring/commit/627b662e6d11b18beaff2266712a59f8845f3d35