@maciej 2018-08-21
Great results!

@e-dard could you, please, share what tool did you use the produce the benchmark result comparison tables?