@e-dard 2018-08-27
Hi @maciej, 

Sorry I only just caught up on this message...

There are two main benchmark comparison tools: [benchcmp](https://godoc.org/golang.org/x/tools/cmd/benchcmp) and [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat).

I prefer `benchstat`.

You need to generate two set of benchmark data (before and after). In order to get enough data to do the U-test you can provide `-count` to the go test tool. For example:

```
$ BENCH_REAL_DATA=true go test -run NONE -bench BenchmarkRealDataFastOr -count 10 -benchmem > old.txt

// Make changes

$ BENCH_REAL_DATA=true go test -run NONE -bench BenchmarkRealDataFastOr -count 10 -benchmem > new.txt
```

You can then generate the results with:

```
$ benchstat old.txt new.txt
```